refactor(core-rooms): group domain under domain/ package
