+++ shared/core-rooms/src/commonMain/kotlin/dev/nonoxy/residetrack/core/rooms/data/mappers/RoomMapper.kt
@@ -3,7 +3,7 @@ package dev.nonoxy.residetrack.core.rooms.data.mappers
 import dev.nonoxy.residetrack.common.utils.mapper.Mapper
 import dev.nonoxy.residetrack.core.database.entities.RoomEntity
 import dev.nonoxy.residetrack.core.database.relations.RoomWithStudents
-import dev.nonoxy.residetrack.core.rooms.models.Room
+import dev.nonoxy.residetrack.core.rooms.domain.model.Room
 
 internal interface RoomMapper : Mapper<RoomWithStudents, Room> {
     fun map(item: Room): RoomEntity
