+++ shared/core-rooms/src/commonMain/kotlin/dev/nonoxy/residetrack/core/rooms/domain/model/Student.kt
@@ -1,4 +1,4 @@
-package dev.nonoxy.residetrack.core.rooms.models
+package dev.nonoxy.residetrack.core.rooms.domain.model
 
 import kotlinx.datetime.LocalDate
