+++ shared/core-rooms/src/commonMain/kotlin/dev/nonoxy/residetrack/core/rooms/domain/upcoming/UpcomingBucket.kt
@@ -1,3 +1,3 @@
-package dev.nonoxy.residetrack.core.rooms.upcoming
+package dev.nonoxy.residetrack.core.rooms.domain.upcoming
 
 enum class UpcomingBucket { OVERDUE, TODAY_TOMORROW, THIS_WEEK }
