+++ shared/feature-room-editor/api/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/room_editor/api/store/RoomEditorStore.kt
@@ -4,7 +4,7 @@ import com.arkivanov.mvikotlin.core.store.Store
 import dev.nonoxy.residetrack.feature.room_editor.api.store.RoomEditorStore.Intent
 import dev.nonoxy.residetrack.feature.room_editor.api.store.RoomEditorStore.Label
 import dev.nonoxy.residetrack.feature.room_editor.api.store.RoomEditorStore.State
-import dev.nonoxy.residetrack.core.rooms.models.Room
+import dev.nonoxy.residetrack.core.rooms.domain.model.Room
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
