+++ shared/feature-rooms/impl/src/commonTest/kotlin/dev/nonoxy/residetrack/feature/rooms/impl/data/FakeRoomsRepository.kt
@@ -1,7 +1,7 @@
 package dev.nonoxy.residetrack.feature.rooms.impl.data
 
-import dev.nonoxy.residetrack.core.rooms.models.Room
-import dev.nonoxy.residetrack.core.rooms.repository.RoomsRepository
+import dev.nonoxy.residetrack.core.rooms.domain.model.Room
+import dev.nonoxy.residetrack.core.rooms.domain.repository.RoomsRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.flow
