+++ shared/feature-rooms/impl/src/commonTest/kotlin/dev/nonoxy/residetrack/feature/rooms/impl/domain/RoomsExecutorTest.kt
@@ -2,7 +2,7 @@ package dev.nonoxy.residetrack.feature.rooms.impl.domain
 
 import com.arkivanov.mvikotlin.extensions.coroutines.labels
 import com.arkivanov.mvikotlin.main.store.DefaultStoreFactory
-import dev.nonoxy.residetrack.core.rooms.models.Room
+import dev.nonoxy.residetrack.core.rooms.domain.model.Room
 import dev.nonoxy.residetrack.feature.rooms.api.store.RoomsStore
 import dev.nonoxy.residetrack.feature.rooms.api.store.RoomsStore.Intent
 import dev.nonoxy.residetrack.feature.rooms.api.store.RoomsStore.Label
