+++ shared/feature-upcoming/api/src/commonMain/kotlin/dev/nonoxy/residetrack/feature/upcoming/api/models/UpcomingItem.kt
@@ -1,6 +1,6 @@
 package dev.nonoxy.residetrack.feature.upcoming.api.models
 
-import dev.nonoxy.residetrack.core.rooms.upcoming.UpcomingBucket
+import dev.nonoxy.residetrack.core.rooms.domain.upcoming.UpcomingBucket
 import kotlinx.datetime.LocalDate
 
 data class UpcomingItem(
